+++ src/classes/babies/ArcadeBaby.ts
@@ -3,8 +3,16 @@ import { Callback } from "isaacscript-common";
 import { getBabyPlayerFromEntity } from "../../utils";
 import { Baby } from "../Baby";
 
+const v = {
+  run: {
+    numTearsFired: 0,
+  },
+};
+
 /** Razor blade tears. */
 export class ArcadeBaby extends Baby {
+  v = v;
+
   @Callback(ModCallback.POST_FIRE_TEAR)
   postFireTear(tear: EntityTear): void {
     const player = getBabyPlayerFromEntity(tear);
@@ -12,8 +20,14 @@ export class ArcadeBaby extends Baby {
       return;
     }
 
-    // Changing the variant does not actually increase the damage, only the appearance.
-    tear.ChangeVariant(TearVariant.RAZOR);
-    tear.CollisionDamage = player.Damage * 3;
+    const num = this.getAttribute("num");
+    v.run.numTearsFired++;
+    if (v.run.numTearsFired === num) {
+      v.run.numTearsFired = 0;
+
+      // Changing the variant does not actually increase the damage, only the appearance.
+      tear.ChangeVariant(TearVariant.RAZOR);
+      tear.CollisionDamage = player.Damage * 3;
+    }
   }
 }